Closer to the end. We now output Prometheus fields via ToString().
