+++ src/include/classes-ea/PrometheusMetrics.h
@@ -91,362 +91,10 @@
 // Includes.
 #include "../classes/File.mqh"
 #include "../classes/Std.h"
-
-/**
- * Stores information about Prometheus schema group.
- */
-class PrometheusSchemaField : public Dynamic {
- public:
-  /* Public fields */
-
-  // Name of the field.
-  string name;
-
-  // Type of the field.
-  ENUM_DATATYPE type;
-
-  // Description of the field.
-  string help;
-
-  /**
-   * Constructor.
-   */
-  PrometheusSchemaField(string _name, ENUM_DATATYPE _type, string _help = "") {
-    name = _name;
-    type = _type;
-    help = _help;
-  }
-};
-
-/**
- * Stores information about Prometheus schema group.
- */
-class PrometheusSchemaGroup : public Dynamic {
- public:
-  /* Protected fields */
-
-  // Name of the group.
-  string name;
-
-  // Type of the group: counter, gauge, histogram, summary.
-  string type;
-
-  // Description of the group.
-  string help;
-
-  // Dict of field name => field information.
-  DictStruct<string, Ref<PrometheusSchemaField>> fields;
-
-  /**
-   * Constructor.
-   */
-  PrometheusSchemaGroup(string _name, string _type, string _help) {
-    name = _name;
-    type = _type;
-    help = _help;
-  }
-
-  /* Public methods */
-
-  /**
-   * Adds field to this schema group.
-   */
-  void AddField(string _field_name, ENUM_DATATYPE _field_type, string _field_help = "") {
-    if (!fields.KeyExists(_field_name)) {
-      Alert("Field \"", _field_name, "\" has been already added to the group \"", name, "\"!");
-      DebugBreak();
-      return;
-    }
-    Ref<PrometheusSchemaField> _field = new PrometheusSchemaField(_field_name, _field_type, _field_help);
-    fields.Set(_field_name, _field);
-  }
-};
-
-/**
- * Class to create schema for Prometheus Metrics to be exported from EA.
- */
-class PrometheusSchema : public Dynamic {
- protected:
-  /* Protected fields */
-
-  // Dict of group name => group information.
-  DictStruct<string, Ref<PrometheusSchemaGroup>> groups;
-
- public:
-  /* Public methods */
-
-  /**
-   * Adds field group to the schema.
-   */
-  void AddGroup(string name, string type, string help = "") {
-    Ref<PrometheusSchemaGroup> _group = new PrometheusSchemaGroup(name, type, help);
-    groups.Set(name, _group);
-  }
-
-  /**
-   * Adds field to the given schema group.
-   */
-  void AddField(string _group_name, string _field_name, ENUM_DATATYPE _field_type, string _field_help = "") {
-    // Searching for the group.
-    unsigned int _group_pos;
-    if (!groups.KeyExists(_group_name, _group_pos)) {
-      Alert("You have to add the group \"", _group_name, "\" via AddGroup() before using AddField()!");
-      DebugBreak();
-      return;
-    }
-    Ref<PrometheusSchemaGroup> _group = groups.GetByPos(_group_pos);
-    _group.Ptr().AddField(_field_name, _field_type, _field_help);
-  }
-};
-
-union TypedValueUnion {
-  bool vbool;
-  long vlong;
-  double vdouble;
-};
-
-enum ENUM_TYPED_VALUE_TYPE {
-  TYPED_VALUE_TYPE_BOOL,
-  TYPED_VALUE_TYPE_INT,
-  TYPED_VALUE_TYPE_LONG,
-  TYPED_VALUE_TYPE_DOUBLE,
-  TYPED_VALUE_TYPE_STRING
-};
-
-/**
- * Class stores a value that can set and retrieved by any type.
- */
-class TypedValue {
- protected:
-  /* Protected fields */
-
-  // POD value stored.
-  TypedValueUnion value;
-
-  // String value stored.
-  string vstring;
-
-  // Date and time value stored.
-  datetime vdatetime;
-
-  // Type of value stored.
-  ENUM_DATATYPE type;
-
- public:
-  /**
-   * Constructor
-   */
-  TypedValue(ENUM_DATATYPE _type) {
-    type = _type;
-
-    switch (type) {
-      case TYPE_BOOL:
-        value.vbool = false;
-        break;
-      case TYPE_CHAR:
-        vstring = "";
-        break;
-      case TYPE_UCHAR:
-      case TYPE_SHORT:
-      case TYPE_USHORT:
-      case TYPE_INT:
-      case TYPE_UINT:
-      case TYPE_LONG:
-      case TYPE_ULONG:
-        value.vlong = 0;
-        break;
-      case TYPE_FLOAT:
-      case TYPE_DOUBLE:
-        value.vdouble = 0;
-        break;
-      case TYPE_STRING:
-        vstring = "";
-        break;
-      case TYPE_DATETIME:
-        vdatetime = (datetime)0;
-        break;
-      default:
-        Alert("Unsupported type: ", EnumToString(type));
-        DebugBreak();
-    }
-  }
-
-  /* Public methods */
-
-  /**
-   * Sets value from given boolean.
-   */
-  void Set(bool _value) {
-    switch (type) {
-      case TYPE_BOOL:
-        value.vbool = _value;
-        break;
-      case TYPE_CHAR:
-        vstring = _value ? "+" : "-";
-        break;
-      case TYPE_UCHAR:
-      case TYPE_SHORT:
-      case TYPE_USHORT:
-      case TYPE_INT:
-      case TYPE_UINT:
-      case TYPE_LONG:
-      case TYPE_ULONG:
-        value.vlong = _value;
-        break;
-      case TYPE_FLOAT:
-      case TYPE_DOUBLE:
-        value.vdouble = _value ? 1.0 : 0;
-        break;
-      case TYPE_STRING:
-        vstring = _value ? "True" : "False";
-        break;
-      case TYPE_DATETIME:
-        Alert("Could not initialize datetime property from bool value!");
-        DebugBreak();
-        break;
-      default:
-        Alert("Unsupported type: ", EnumToString(type));
-        DebugBreak();
-    }
-  }
-
-  /**
-   * Sets value from given number.
-   */
-  void Set(long _value) {
-    switch (type) {
-      case TYPE_BOOL:
-        value.vbool = _value != 0;
-        break;
-      case TYPE_CHAR:
-        vstring = (_value != 0) ? "+" : "-";
-        break;
-      case TYPE_UCHAR:
-      case TYPE_SHORT:
-      case TYPE_USHORT:
-      case TYPE_INT:
-      case TYPE_UINT:
-      case TYPE_LONG:
-      case TYPE_ULONG:
-        value.vlong = _value;
-        break;
-      case TYPE_FLOAT:
-      case TYPE_DOUBLE:
-        value.vdouble = (double)_value;
-        break;
-      case TYPE_STRING:
-        vstring = IntegerToString(_value);
-        break;
-      case TYPE_DATETIME:
-        vdatetime = (datetime)_value;
-        break;
-      default:
-        Alert("Unsupported type: ", EnumToString(type));
-        DebugBreak();
-    }
-  }
-
-  /**
-   * Sets value from given number.
-   */
-  void Set(double _value) {
-    switch (type) {
-      case TYPE_BOOL:
-        value.vbool = _value != 0;
-        break;
-      case TYPE_CHAR:
-        vstring = (_value != 0) ? "+" : "-";
-        break;
-      case TYPE_UCHAR:
-      case TYPE_SHORT:
-      case TYPE_USHORT:
-      case TYPE_INT:
-      case TYPE_UINT:
-      case TYPE_LONG:
-      case TYPE_ULONG:
-        value.vlong = (long)_value;
-        break;
-      case TYPE_FLOAT:
-      case TYPE_DOUBLE:
-        value.vdouble = _value;
-        break;
-      case TYPE_STRING:
-        vstring = DoubleToString(_value);
-        break;
-      case TYPE_DATETIME:
-        Alert("Could not initialize datetime property from double value!");
-        DebugBreak();
-        break;
-      default:
-        Alert("Unsupported type: ", EnumToString(type));
-        DebugBreak();
-    }
-  }
-
-  /**
-   * Sets value from given string.
-   */
-  void Set(string _value) {
-    switch (type) {
-      case TYPE_BOOL:
-        value.vbool = _value == "+" || _value == "True" || _value == "true" || _value == "Yes" || _value == "yes" ||
-                      _value == "On" || _value == "on" || _value == "Enabled" || _value == "enabled";
-        break;
-      case TYPE_CHAR:
-        vstring = StringSubstr(_value, 0, 1);
-        break;
-      case TYPE_UCHAR:
-      case TYPE_SHORT:
-      case TYPE_USHORT:
-      case TYPE_INT:
-      case TYPE_UINT:
-      case TYPE_LONG:
-      case TYPE_ULONG:
-        value.vlong = StringToInteger(_value);
-        break;
-      case TYPE_FLOAT:
-      case TYPE_DOUBLE:
-        value.vdouble = StringToDouble(_value);
-        break;
-      case TYPE_STRING:
-        vstring = _value;
-        break;
-      case TYPE_DATETIME:
-        vdatetime = StringToTime(_value);
-        break;
-      default:
-        Alert("Unsupported type: ", EnumToString(type));
-        DebugBreak();
-    }
-  }
-};
-
-/**
- * Structure stores value for the Prometheus metrics field.
- */
-struct PrometheusMetricsValue {
-  // Name of the group.
-  string group_name;
-
-  // Name of the field.
-  string field_name;
-
-  // Value for the field.
-  TypedValue value;
-
-  /**
-   * Default constructor.
-   */
-  PrometheusMetricsValue() : value(TYPE_INT) {}
-
-  /**
-   * Constructor.
-   */
-  PrometheusMetricsValue(string _group_name, string _field_name, ENUM_DATATYPE _field_type) : value(_field_type) {
-    group_name = _group_name;
-    field_name = _field_name;
-  }
-};
+#include "PrometheusMetricsValue.h"
+#include "PrometheusSchema.h"
+#include "PrometheusSchemaField.h"
+#include "PrometheusSchemaGroup.h"
 
 /**
  * Class to export Prometheus Metrics from EA.
@@ -472,7 +120,10 @@ class PrometheusMetrics : public Dynamic {
   /**
    * Class constructor.
    */
-  PrometheusMetrics() { Init(); }
+  PrometheusMetrics(PrometheusSchema* _schema = NULL) {
+    schema = _schema;
+    Init();
+  }
 
   /**
    * Class deconstructor.
@@ -484,31 +135,97 @@ class PrometheusMetrics : public Dynamic {
   /**
    * Prepares for a new set of values.
    */
-  void Clear() {
-    // @todo
-  }
+  void Clear() { values.Clear(); }
 
   /**
    * Outputs values as Prometheus-compatible code.
    */
   string ToString() {
-    // @todo
-    return "";
+    if (!schema.IsSet()) {
+      Alert("Schema was not set! Please use PrometheusMetrics::SetSchema().");
+      DebugBreak();
+      return "<missing schema>";
+    }
+
+    string _out;
+    // Iterating over schema groups, then fields.
+    for (DictStructIterator<string, Ref<PrometheusSchemaGroup>> iter = schema.Ptr().GetGroups().Begin(); iter.IsValid();
+         ++iter) {
+      PrometheusSchemaGroup* _group = iter.Value().Ptr();
+
+      if (iter.Index() != 0) {
+        // Group separator.
+        _out += "\n";
+      }
+
+      if (StringLen(_group.help) > 0) {
+        _out += "HELP " + _group.help + "\n";
+      }
+
+      _out += "# TYPE " + _group.name + " " + _group.type + "\n";
+
+      // Iterating over group fields.
+      for (DictStructIterator<string, Ref<PrometheusSchemaField>> iter2 = _group.GetFields().Begin(); iter2.IsValid();
+           ++iter2) {
+        PrometheusSchemaField* _field = iter2.Value().Ptr();
+        PrometheusMetricsValue _value;
+        if (!TryGetValue(_group.name, _field.name, _value)) {
+          continue;
+        }
+
+        _out += _field.name + " " + _value.ToString() + "\n";
+      }
+    }
+
+    return _out;
+  }
+
+  /**
+   * Tries to get value for the given field.
+   */
+  bool TryGetValue(string _group_name, string _field_name, PrometheusMetricsValue& out) {
+    unsigned int _value_pos;
+    if (values.KeyExists(_group_name + ":" + _field_name, _value_pos)) {
+      out = values.GetByPos(_value_pos);
+      return true;
+    }
+    return false;
   }
 
   /**
    * Checks if we can set given field to given type of value.
    */
   ENUM_DATATYPE CheckSchemaFieldCompatibility(string _group_name, string _field_name, ENUM_DATATYPE _type) {
+    if (!schema.IsSet()) {
+      Alert("Schema was not set! Please use PrometheusMetrics::SetSchema().");
+      DebugBreak();
+      return (ENUM_DATATYPE)-1;
+    }
+
     // @todo
-    return TYPE_INT;
+    return _type;
+  }
+
+  /**
+   * Returns given schema field.
+   */
+  PrometheusSchemaField* GetSchemaField(string _group_name, string _field_name) {
+    if (!schema.IsSet()) {
+      Alert("Schema was not set! Please use PrometheusMetrics::SetSchema().");
+      DebugBreak();
+      return NULL;
+    }
+
+    return schema.Ptr().GetField(_group_name, _field_name);
   }
 
   /**
    * Sets value of the given field.
    */
-  void SetValue(string _group_name, string _field_name, bool value) {
-    ENUM_DATATYPE _field_type = CheckSchemaFieldCompatibility(_group_name, _field_name, TYPE_BOOL);
+  template <typename T>
+  void SetValue(string _group_name, string _field_name, T value) {
+    T _typed_value = T();
+    ENUM_DATATYPE _field_type = CheckSchemaFieldCompatibility(_group_name, _field_name, GetType(_typed_value));
 
     PrometheusMetricsValue _value(_group_name, _field_name, _field_type);
     _value.value.Set(value);
